migrate to webflux + r2dbc consider as pre-alpha and untested

+++ src/main/kotlin/com/young_zy/forum/controller/request/GiveSystemAdminRequest.kt
@@ -1,5 +1,5 @@
 package com.young_zy.forum.controller.request
 
 data class GiveSystemAdminRequest(
-        val userIds: List<Int>
+        val userIds: List<Long>
 )
